pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes"
version = "0.1.0"
description = "Hermes is an AI workflow prototype for sales enablement in experience commerce and location-based entertainment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pandas>=2.2",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "openpyxl>=3.1",
]

[tool.setuptools]
packages = ["app"]
